% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeptime-package.R
\docType{package}
\name{deeptime-package}
\alias{deeptime}
\alias{deeptime-package}
\title{deeptime: Plotting Tools for Anyone Working in Deep Time}
\description{
Extends the functionality of other plotting packages like 'ggplot2' and 'lattice' to help facilitate the plotting of data over long time intervals, including, but not limited to, geological, evolutionary, and ecological data. The primary goal of 'deeptime' is to enable users to add highly customizable timescales to their visualizations. Other functions are also included to assist with other areas of deep time visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/willgearty/deeptime}
  \item \url{https://williamgearty.com/deeptime/}
  \item Report bugs at \url{https://github.com/willgearty/deeptime/issues}
}

}
\author{
\strong{Maintainer}: William Gearty \email{willgearty@gmail.com}

}
\keyword{internal}
