% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{geo_pattern}
\alias{geo_pattern}
\alias{geo_grob}
\title{Get a FGDC geologic plotting pattern}
\usage{
geo_pattern(
  code,
  scale = 2,
  col = NULL,
  fill = NULL,
  alpha = NULL,
  bg = "white"
)

geo_grob(code, col = NULL, fill = NULL, alpha = NULL, bg = "white")
}
\arguments{
\item{code}{The number corresponding to the pattern to return. Strings and
numbers are permitted. See the "pattern numbers" in the \href{https://ngmdb.usgs.gov/fgdc_gds/geolsymstd/fgdc-geolsym-patternchart.pdf}{full pattern chart}
for valid \code{code} values.}

\item{scale}{The visual scale of the pattern.}

\item{col}{The color to use for the lines of the pattern.}

\item{fill}{The color with which to fill the pattern.}

\item{alpha}{The transparency to use for the fill of the pattern.}

\item{bg}{The background color to use for the pattern.}
}
\value{
\code{geo_grob()} returns a \link[grid:grid-defunct]{grob} object with a single
instance of the desired pattern. \code{geo_pattern()} returns a
\link[grid:patterns]{GridPattern} object with a repeated instance of the
desired pattern.
}
\description{
Retrieve a single geologic pattern as defined in the \href{https://ngmdb.usgs.gov/fgdc_gds/geolsymstd/download.php}{FGDC Digital Cartographic Standard for Geologic Map Symbolization} by
the \href{https://ngmdb.usgs.gov/fgdc_gds/index.php}{Geologic Data Subcommittee (GDS)} of the \href{https://www.fgdc.gov/}{Federal Geographic Data Committee (FGDC)}.
}
\details{
These patterns were originally processed and optimized by Daven
Quinn and are hosted on
\href{https://github.com/davenquinn/geologic-patterns/}{GitHub}.
}
\examples{
library(grid)
# Get a generic igneous pattern
pattern1 <- geo_pattern(code = "313-K")
# Get the pattern for a sandstone
pattern2 <- geo_pattern(code = "607")

# plot the two patterns
grid.newpage()
grid.draw(rectGrob(gp = gpar(fill = pattern1)))
grid.newpage()
grid.draw(rectGrob(gp = gpar(fill = pattern2)))
}
\seealso{
FGDC patterns:
\code{\link{grid.pattern_geo}()},
\code{\link{scale_fill_geopattern}()}
}
\concept{patterns}
