% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{periods}
\alias{periods}
\title{Period data from the International Commission on Stratigraphy (v2021/05)}
\format{
A data frame with 22 rows and 5 variables:
\describe{
  \item{name}{period name}
  \item{max_age}{maximum age, in millions of years}
  \item{min_age}{minimum age, in millions of years}
  \item{abbr}{period name abbreviations}
  \item{color}{the colors for each period, according to the Commission for the Geological Map of the World}
}
}
\source{
\url{https://stratigraphy.org} via \url{https://macrostrat.org/api/v2/defs/intervals?timescale=international\%20periods}
}
\usage{
periods
}
\description{
A dataset containing the boundary ages, abbreviations, and colors for the periods of the Geologic Time Scale.
Based on The ICS International Chronostratigraphic Chart (v2021/05), by Cohen, Finney, Gibbard, and Fan.
}
\keyword{datasets}
