% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{scale_fill_geopattern}
\alias{scale_fill_geopattern}
\title{Geologic pattern fill scale}
\usage{
scale_fill_geopattern(na.value = "grey50", ...)
}
\arguments{
\item{na.value}{The aesthetic value to use for missing (NA) values. May be
either a color or a \link[grid:patterns]{GridPattern} object (such as that
returned by \code{\link[=geo_pattern]{geo_pattern()}}).}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale}}
  \describe{
    \item{\code{scale_name}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The name of the scale
that should be used for error messages associated with this scale.}
    \item{\code{name}}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}
    \item{\code{breaks}}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{labels}}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item An expression vector (must be the same length as breaks). See ?plotmath for details.
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{limits}}{One of:
\itemize{
\item \code{NULL} to use the default scale values
\item A character vector that defines possible values of the scale and their
order
\item A function that accepts the existing (automatic) values and returns
new ones. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{na.translate}}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}
    \item{\code{drop}}{Should unused factor levels be omitted from the scale?
The default, \code{TRUE}, uses the levels that appear in the data;
\code{FALSE} includes the levels in the factor. Please note that to display
every level in a legend, the layer should use \code{show.legend = TRUE}.}
    \item{\code{guide}}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}
    \item{\code{call}}{The \code{call} used to construct the scale for reporting messages.}
  }}
}
\description{
Fill scale using the \href{https://ngmdb.usgs.gov/fgdc_gds/geolsymstd/download.php}{FGDC Digital Cartographic Standard for Geologic Map Symbolization}. Fill
values should correspond to the "pattern numbers" in the \href{https://ngmdb.usgs.gov/fgdc_gds/geolsymstd/fgdc-geolsym-patternchart.pdf}{full pattern chart}.
See \code{\link[=geo_pattern]{geo_pattern()}} for more details. Note that \code{ggplot2} >= 3.5.0 is
required to use this function.
}
\section{Warning}{
 Pattern fills are not supported on all graphics devices.
Where they are not supported, closed shapes will be rendered with a
transparent fill. For example, on Windows machines, the default device in
RStudio and in the knitr package is \code{\link[=png]{png()}}, which does not support
patterns. In RStudio, you can go to ‘Tools > Global Options > General >
Graphics’ and choose the ‘Cairo PNG’ device from the dropdown
menu to display patterns.
}

\examples{
\dontshow{if (packageVersion("ggplot2") >= "3.5.0") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggplot2)
vals <- c("101", "313", "603", "733")
ggplot(mpg, aes(factor(cyl), fill = vals[factor(cyl)])) +
  geom_bar() +
  scale_fill_geopattern(name = NULL)
\dontshow{\}) # examplesIf}
}
\seealso{
FGDC patterns:
\code{\link{geo_pattern}()},
\code{\link{grid.pattern_geo}()}
}
\concept{patterns}
