% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stages}
\alias{stages}
\title{Stage data based on The Geologic Time Scale 2012}
\format{A data frame with 15 rows and 5 variables:
\describe{
  \item{name}{stage name}
  \item{max_age}{maximum age, in millions of years}
  \item{min_age}{minimum age, in millions of years}
  \item{color}{the colors for each stage, according to the Comission for the Geological Map of the World}
}}
\source{
\url{https://doi.org/10.1016/C2011-1-08249-8}
}
\usage{
stages
}
\description{
A dataset containing the boundary ages, abbrevations, and colors for the stages of the Geologic Time Scale.
Based on The Geologic Time Scale 2012, by Gradstein and Ogg.
}
\keyword{datasets}
