% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epochs}
\alias{epochs}
\title{Epoch data from the International Commission on Stratigraphy (v2022/10)}
\format{
A data frame with 34 rows and 5 variables:
\describe{
\item{name}{epoch name}
\item{max_age}{maximum age, in millions of years}
\item{min_age}{minimum age, in millions of years}
\item{abbr}{epoch name abbreviations}
\item{color}{the colors for each epoch, according to the Commission for the Geological Map of the World}
}
}
\source{
\url{https://stratigraphy.org} via \url{https://macrostrat.org/api/v2/defs/intervals?timescale=international\%20epochs}
}
\usage{
epochs
}
\description{
A dataset containing the boundary ages, abbreviations, and colors for the epochs of the Geologic Time Scale.
Based on The ICS International Chronostratigraphic Chart (v2022/10), by Cohen, Finney, Gibbard, and Fan.
}
\seealso{
Other timescales: 
\code{\link{eons}},
\code{\link{eras}},
\code{\link{periods}},
\code{\link{stages}}
}
\concept{timescales}
\keyword{datasets}
