% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scale_data.R
\name{getScaleData}
\alias{getScaleData}
\title{Get geological timescale data}
\usage{
getScaleData(name)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{getScaleData()} was renamed to \code{\link[=get_scale_data]{get_scale_data()}} as of \strong{deeptime} version
1.0.0 to create a more consistent API.
}
\keyword{internal}
