% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggeo_scale.R
\name{gggeo_scale}
\alias{gggeo_scale}
\title{Add a geologic scale below ggplots using \emph{grid}}
\usage{
gggeo_scale(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{ggplot}, \code{gtable}, or \code{geo_scale} (as produced
by this function).}
}
\value{
A geo_scale object. Basically a gtable object but with the axis
limits included.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function takes a ggplot object and adds a geologic time scale at the
specified side using the grid and gtable packages.

If custom data is provided (with \code{dat}), it should consist of at least 3
columns of data. See \code{data(periods)} for an example.
\itemize{
\item The \code{name} column lists the names of each time interval. These will
be used as labels if no abbreviations are provided.
\item The \code{max_age} column lists the oldest boundary of each time interval.
\item The \code{min_age} column lists the youngest boundary of each time
interval.
\item The \code{abbr} column is optional and lists abbreviations that may be
used as labels.
\item The \code{color} column is also optional and lists a hex color code
(which can be obtained with \code{rgb()}) for each time interval.
}
}
\section{Life cycle}{

This function is fully deprecated in favor of \code{\link[=coord_geo]{coord_geo()}} as of
\strong{deeptime} version 2.3.0. It will be removed in a future version.
}

\keyword{internal}
