% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggeo_scale.R
\name{gggeo_scale}
\alias{gggeo_scale}
\alias{gggeo_scale.gtable}
\alias{gggeo_scale.ggplot}
\alias{gggeo_scale.geo_scale}
\alias{print.geo_scale}
\title{Add a geologic scale to ggplots}
\usage{
gggeo_scale(x, ...)

\method{gggeo_scale}{gtable}(gt, lims, dat = "periods", fill = NULL,
  color = "black", alpha = 1, height = unit(2, "line"),
  pos = "bottom", lab = TRUE, rot = 0, abbrv = TRUE,
  skip = c("Quaternary", "Holocene", "Late Pleistocene"), size = 5,
  margin = unit(0.5, "line"), neg = FALSE)

\method{gggeo_scale}{ggplot}(gg, dat = "periods", fill = NULL,
  color = "black", alpha = 1, height = unit(2, "line"),
  pos = "bottom", lab = TRUE, rot = 0, abbrv = TRUE,
  skip = c("Quaternary", "Holocene", "Late Pleistocene"), size = 5,
  margin = unit(0.5, "line"), neg = FALSE)

\method{gggeo_scale}{geo_scale}(geo, dat = "periods", fill = NULL,
  color = "black", alpha = 1, height = unit(2, "line"),
  pos = "bottom", lab = TRUE, rot = 0, abbrv = TRUE,
  skip = c("Quaternary", "Holocene", "Late Pleistocene"), size = 5,
  margin = unit(0.5, "line"), neg = FALSE)

\method{print}{geo_scale}(geo, ...)
}
\arguments{
\item{x}{Either a ggplot object or a gtable object.}

\item{...}{further arguments passed to \code{grid.draw}.}

\item{gt}{A gtable object.}

\item{lims}{The limits of the axis of the desired side of the plot. Only required if using a gtable object not created by this function.}

\item{dat}{Either A) a string indicating a built-in dataframe with interval data from the ICS ("periods", "epochs", "stages", "eons", or "eras"),
B) a string indicating a timescale from macrostrat (see list here: \url{https://macrostrat.org/api/defs/timescales?all}),
or C) a custom dataframe of time interval boundaries (see Details).}

\item{fill}{The fill color of the boxes. The default is to use the colors included in \code{dat}.
If a custom dataset is provided with \code{dat} without color and without fill, a greyscale will be used.
Custom fill colors can be provided with this option and will be recycled if/as necessary.}

\item{color}{The outline color of the interval boxes.}

\item{alpha}{The transparency of the fill colors.}

\item{height}{The height (or width if \code{pos} is \code{left} or \code{right}) of the scale.}

\item{pos}{Which side to add the scale to (left, right, top, or bottom). First letter may also be used.}

\item{lab}{Whether to include labels.}

\item{rot}{The amount of counter-clockwise rotation to add to the labels (in degrees).}

\item{abbrv}{If including labels, whether to use abbreviations instead of full interval names.}

\item{skip}{A vector of interval names indicating which intervals should not be labelled.}

\item{size}{Label size.}

\item{margin}{The width of the margin around the returned object (can be a vector of length 4).}

\item{neg}{Set this to true if your x-axis is using negative values.}

\item{gg}{A ggplot object.}

\item{geo}{A geo_scale object output by \code{gggeo_scale()}.}
}
\value{
A geo_scale object. Basically a gtable object but with the axis limits included.
}
\description{
This function takes a ggplot object and adds a geologic time scale at the specified side.
}
\details{
If custom data is provided (with \code{dat}), it should consist of at least 3 columns of data. See \code{data(periods)} for an example.
  The \code{name} column lists the names of each time interval. These will be used as labels if no abbreviations are provided.
  The \code{max_age} column lists the oldest boundary of each time interval.
  The \code{min_age} column lists the youngest boundary of each time interval.
  The \code{abbr} column is optional and lists abbreviations that may be used as labels.
  The \code{color} column is also optional and lists a hex color code (which can be obtained with \code{rgb()}) for each time interval.
}
\examples{
library(ggplot2)
# bottom scale by default
p <- ggplot() +
  geom_point(aes(y = runif(1000, 0, 8), x = runif(1000, 0, 1000))) +
  scale_x_reverse() +
  coord_cartesian(xlim = c(0, 1000), ylim = c(0,8), expand = FALSE) +
  theme_classic()
gggeo_scale(p)

# can specify any side of the plot
p <- ggplot() +
 geom_point(aes(x = runif(1000, 0, 8), y = runif(1000, 0, 1000))) +
 scale_y_reverse() +
 coord_cartesian(xlim = c(0, 8), ylim = c(0,1000), expand = FALSE) +
 theme_classic()
gggeo_scale(p, pos = "left", rot = 90)

# can add multiple scales
p <- ggplot() +
  geom_point(aes(y = runif(1000, 0, 8), x = runif(1000, 0, 100))) +
  scale_x_reverse() +
  coord_cartesian(xlim = c(0, 100), ylim = c(0,8), expand = FALSE) +
  theme_classic()
p <- gggeo_scale(p, abbrv = FALSE)
p <- gggeo_scale(p, lims = c(-100,0), dat = "epochs", height = unit(4, "lines"), rot = 90,
                 size = 2.5, abbrv = FALSE)
gggeo_scale(p, lims = c(-100,0), dat = "stages", height = unit(4, "lines"), rot = 90, size = 2.5,
            abbrv = FALSE)

# intervals on both sides for different timescales (ICS stages vs North American Land Mammal Ages)
p <- ggplot() +
  geom_point(aes(x = runif(1000, 0, 10), y = runif(1000, 0, 65))) +
  scale_y_reverse() +
  coord_cartesian(xlim = c(0, 10), ylim = c(0,65), expand = FALSE) +
  theme_classic()
p <- gggeo_scale(p, dat = "stages", pos = "left", height = unit(4, "lines"), size = 2.5,
                 abbrv = FALSE)
gggeo_scale(p, dat = "North American Land Mammal Ages", pos = "right", height = unit(4, "lines"),
            size = 2.5, abbrv = FALSE)

#can add scales to a faceted plot
#use gggeo_scale_old() if you have more than one column
df <- data.frame(x = runif(1000, 0, 541), y = runif(1000, 0, 8), z = sample(c(1,2,3,4), 1000, TRUE))
p <- ggplot(df) +
  geom_point(aes(x, y)) +
  scale_x_reverse() +
  coord_cartesian(xlim = c(0, 541), ylim = c(0,8), expand = FALSE) +
  theme_classic() +
  facet_wrap(~z, ncol = 1)
gggeo_scale(p)

#can even add a scale to a phylogeny (using ggtree)
\dontrun{
library(phytools)
library(ggtree)
tree <- pbtree(b = .03, d = .01,  n=100)
p <- ggtree(tree) +
  coord_cartesian(xlim = c(0,-500), ylim = c(-2,Ntip(tree)), expand = FALSE) +
  scale_x_continuous(breaks=seq(-500,0,100), labels=abs(seq(-500,0,100))) +
  theme_tree2()
p <- revts(p)
gggeo_scale(p, neg = TRUE)
}
}
