% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggeo_scale.R
\name{gggeo_scale}
\alias{gggeo_scale}
\title{Add a geologic scale to ggplots}
\usage{
gggeo_scale(gg, fill = NULL, color = "black", alpha = 1,
  height = 0.05, size = 5, quat = FALSE, pos = "bottom",
  abbrv = TRUE, dat = NULL, neg = FALSE)
}
\arguments{
\item{gg}{The ggplot object.}

\item{fill}{The fill color of the boxes. The default is to use the colors included in \code{dat}.
If a custom dataset is provided with dat without color and without fill, a greyscale will be used
Custom fill colors can be provided with this option and will be recycled if/as necessary.}

\item{color}{The outline color of the interval boxes.}

\item{alpha}{The transparency of the fill colors.}

\item{height}{The proportional height (or width if \code{pos} is \code{left} or \code{right}) of the entire plot to use for the scale.}

\item{size}{Label size.}

\item{quat}{Specifies whether the Quaternary should be labelled (if using the default data).}

\item{pos}{Which side to add the scale to (left, right, top, or bottom). First letter may also be used.}

\item{abbrv}{Whether to use abbreviations instead of full interval names.}

\item{dat}{A custom data set of time interval boundaries (see Details).}

\item{neg}{Set this to true if your x-axis is using negative values.}
}
\value{
A ggplot object.
}
\description{
This function takes a ggplot object and adds a geologic time scale at the specified side.
}
\details{
If custom data is provided (with \code{dat}), it should consist of at least 3 columns of data. See \code{data(dat)} for an example.
  The \code{name} column lists the names of each time interval. These will be used as labels if no abbreviations are provided.
  The \code{max_age} column lists the oldest boundary of each time interval.
  The \code{min_age} column lists the youngest boundary of each time interval.
  The \code{abbr} column is optonal and lists abbreviations that may be used as labels.
  The \code{color} column is also optional and lists a hex color code (which can be obtained with \code{rgb()}) for each time interval.
}
\examples{
library(ggplot2)
# bottom scale by default
p <- ggplot() +
  geom_point(aes(y = runif(1000, .5, 8), x = runif(1000, 0, 1000))) +
  scale_x_reverse() +
  coord_cartesian(xlim = c(0, 1000), ylim = c(0,8), expand = FALSE) +
  theme_classic()
gggeo_scale(p)

# can specify any side of the plot
p <- ggplot() +
 geom_point(aes(x = runif(1000, .5, 8), y = runif(1000, 0, 1000))) +
 scale_y_reverse() +
 coord_cartesian(xlim = c(0, 8), ylim = c(0,1000), expand = FALSE) +
 theme_classic()
gggeo_scale(p, pos = "left")

#can even add a scale to a phylogeny (using ggtree)
library(phytools)
library(ggtree)
tree <- pbtree(b = .03, d = .01,  n=100)
p <- ggtree(tree) +
 coord_cartesian(xlim = c(0,-500), ylim = c(-10,Ntip(tree)), expand = FALSE) +
 scale_x_continuous(breaks=seq(-300,0,100), labels=abs(seq(-300,0,100))) +
 theme_tree2()
p <- revts(p)
gggeo_scale(p, neg = TRUE)
}
