% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity_through_time.R
\name{panel.disparity}
\alias{panel.disparity}
\title{Combined wireframe and cloud panel}
\usage{
panel.disparity(x, y, z, groups, subscripts, ...)
}
\arguments{
\item{x, y, z, groups, subscripts, ...}{Same as for \code{\link[lattice:panel.cloud]{lattice::panel.cloud()}}}
}
\value{
No return value, plots the results of both \code{\link[lattice:panel.cloud]{lattice::panel.cloud()}} and \code{\link[lattice:panel.cloud]{lattice::panel.wireframe()}}.
}
\description{
Plots the provided data on 2-D surfaces within a 3-D framework. See \code{\link[=disparity_through_time]{disparity_through_time()}}.
}
